Program to use as external process
This is the program I used to test QProcess. I wanted to test reading
both stdout and stderr messages as well
as seeing how readyReadStandardOutput() deals with sleeping in the app.
My assumption based on results is
that QProcess will only quit reading when the other application starts
to ask for input itself. Otherwise it will
not return EOF just because there is a delay.